+++ packages/adapters-library/src/adapters/pendle/backend/constants.ts
@@ -1,0 +1,6 @@
+export const PENDLE_ORACLE_ADDRESS_ALL_CHAINS =
+  '0x9a9fa8338dd5e5b2188006f1cd2ef26d921650c2'
+
+// I don't know what these values mean - JP
+export const DURATION_15_MINS = 900
+export const DURATION_30_MINS = 1800
