--- packages/adapters-library/src/adapters/pendle/products/market/metadata/optimism.market.json
@@ -1,130 +1,0 @@
-{
-  "0xF4225F061E5E01aA59DE5E615729A9180301EB07": {
-    "protocolToken": {
-      "address": "0xF4225F061E5E01aA59DE5E615729A9180301EB07",
-      "name": "PT wstETH 26SEP2024",
-      "symbol": "PT-wstETH-26SEP2024",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x96A528f4414aC3CcD21342996c93f2EcdEc24286",
-      "name": "SY wstETH",
-      "symbol": "SY-wstETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x24fB77c8C776C75F869BC65e6856AF56f57d919F",
-    "type": "principleToken"
-  },
-  "0x1DFe41cc7F7860BA7f1076ca6d0fedD707c87A00": {
-    "protocolToken": {
-      "address": "0x1DFe41cc7F7860BA7f1076ca6d0fedD707c87A00",
-      "name": "YT wstETH 26SEP2024",
-      "symbol": "YT-wstETH-26SEP2024",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x96A528f4414aC3CcD21342996c93f2EcdEc24286",
-      "name": "SY wstETH",
-      "symbol": "SY-wstETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x24fB77c8C776C75F869BC65e6856AF56f57d919F",
-    "type": "yieldToken"
-  },
-  "0x96A528f4414aC3CcD21342996c93f2EcdEc24286": {
-    "protocolToken": {
-      "address": "0x96A528f4414aC3CcD21342996c93f2EcdEc24286",
-      "name": "SY wstETH",
-      "symbol": "SY-wstETH",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x1F32b1c2345538c0c6f582fCB022739c4A194Ebb",
-      "name": "Wrapped liquid staked Ether 2.0",
-      "symbol": "wstETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x24fB77c8C776C75F869BC65e6856AF56f57d919F",
-    "type": "standardisedYieldToken"
-  },
-  "0x24fB77c8C776C75F869BC65e6856AF56f57d919F": {
-    "protocolToken": {
-      "address": "0x24fB77c8C776C75F869BC65e6856AF56f57d919F",
-      "name": "Pendle Market",
-      "symbol": "PENDLE-LPT",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x1F32b1c2345538c0c6f582fCB022739c4A194Ebb",
-      "name": "Wrapped liquid staked Ether 2.0",
-      "symbol": "wstETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x24fB77c8C776C75F869BC65e6856AF56f57d919F",
-    "type": "lpToken"
-  },
-  "0xAA92Ea52612fA5ee5aba5510682485c19750CB82": {
-    "protocolToken": {
-      "address": "0xAA92Ea52612fA5ee5aba5510682485c19750CB82",
-      "name": "PT rETH 27JUN2024",
-      "symbol": "PT-rETH-27JUN2024",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x9D6d509C0354ACA187aAC6beA7d063D3Ef68e2A0",
-      "name": "SY rETH",
-      "symbol": "SY-rETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x0C485Feb9E6fee816652Ea8f3beD2A8f59296E40",
-    "type": "principleToken"
-  },
-  "0xE8F0cf61598E0d7D20f99b978BdeFF4D4222FeB8": {
-    "protocolToken": {
-      "address": "0xE8F0cf61598E0d7D20f99b978BdeFF4D4222FeB8",
-      "name": "YT rETH 27JUN2024",
-      "symbol": "YT-rETH-27JUN2024",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x9D6d509C0354ACA187aAC6beA7d063D3Ef68e2A0",
-      "name": "SY rETH",
-      "symbol": "SY-rETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x0C485Feb9E6fee816652Ea8f3beD2A8f59296E40",
-    "type": "yieldToken"
-  },
-  "0x9D6d509C0354ACA187aAC6beA7d063D3Ef68e2A0": {
-    "protocolToken": {
-      "address": "0x9D6d509C0354ACA187aAC6beA7d063D3Ef68e2A0",
-      "name": "SY rETH",
-      "symbol": "SY-rETH",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x9Bcef72be871e61ED4fBbc7630889beE758eb81D",
-      "name": "Rocket Pool ETH",
-      "symbol": "rETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x0C485Feb9E6fee816652Ea8f3beD2A8f59296E40",
-    "type": "standardisedYieldToken"
-  },
-  "0x0C485Feb9E6fee816652Ea8f3beD2A8f59296E40": {
-    "protocolToken": {
-      "address": "0x0C485Feb9E6fee816652Ea8f3beD2A8f59296E40",
-      "name": "Pendle Market",
-      "symbol": "PENDLE-LPT",
-      "decimals": 18
-    },
-    "underlyingToken": {
-      "address": "0x9Bcef72be871e61ED4fBbc7630889beE758eb81D",
-      "name": "Rocket Pool ETH",
-      "symbol": "rETH",
-      "decimals": 18
-    },
-    "marketAddress": "0x0C485Feb9E6fee816652Ea8f3beD2A8f59296E40",
-    "type": "lpToken"
-  }
-}
