+++ packages/adapters-library/src/adapters/pendle/products/principle-token/metadata/bsc.market.json
@@ -1,0 +1,87 @@
+{
+  "0x70c1138B54ba212776d3A9d29b6160C54C31cd5d": {
+    "protocolToken": {
+      "address": "0x70c1138B54ba212776d3A9d29b6160C54C31cd5d",
+      "name": "PT wbETH 26DEC2024",
+      "symbol": "PT-wbETH-26DEC2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x96F77F24b3d17166FcaF3a9E3B8330fF38B966F9",
+        "name": "SY wbETH",
+        "symbol": "SY-wbETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x080f52A881ba96EEE2268682733C857c560e5dd4"
+  },
+  "0x5eC2ae0AFDEc891E7702344dc2A31C636B3627Eb": {
+    "protocolToken": {
+      "address": "0x5eC2ae0AFDEc891E7702344dc2A31C636B3627Eb",
+      "name": "PT Thena frxETH-ETH 27JUN2024",
+      "symbol": "PT-Thena frxETH-ETH-27JUN2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0xf86DDA825885674F85C5420079E1Df1bC0F372c6",
+        "name": "SY Thena frxETH-ETH",
+        "symbol": "SY-Thena frxETH-ETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x66317Dec4a3d8d1e47b85f704E5DF675a68BB7C9"
+  },
+  "0x04eb6B56ff53f457c8E857ca8D4fbC8d9a531c0C": {
+    "protocolToken": {
+      "address": "0x04eb6B56ff53f457c8E857ca8D4fbC8d9a531c0C",
+      "name": "PT ankrBNB 26JUN2025",
+      "symbol": "PT-ankrBNB-26JUN2025",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x58732F1C1243A0D5876a601FB8D1Ee508B21580D",
+        "name": "SY ankrBNB",
+        "symbol": "SY-ankrBNB",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0xEDa1D0e1681D59dea451702963d6287b844Cb94C"
+  },
+  "0x5eC5033a44ffb6f21DA3A4f4CCe4163485D61747": {
+    "protocolToken": {
+      "address": "0x5eC5033a44ffb6f21DA3A4f4CCe4163485D61747",
+      "name": "PT Renzo ezETH 26SEP2024",
+      "symbol": "PT-ezETH-26SEP2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0xe49269B5D31299BcE407c8CcCf241274e9A93C9A",
+        "name": "SY Renzo ezETH",
+        "symbol": "SY-ezETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0xB2eEA27aF50030A445418553c4892065cf3a720a"
+  },
+  "0x0E8Ba474FCB6581059C61c328a21CFa285C7Cee7": {
+    "protocolToken": {
+      "address": "0x0E8Ba474FCB6581059C61c328a21CFa285C7Cee7",
+      "name": "PT weETH 26SEP2024",
+      "symbol": "PT-weETH-26SEP2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x56b70AC3a6Ae850F6AEE67591d1bB86895Ba211E",
+        "name": "SY weETH",
+        "symbol": "SY-weETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x27255F9aFF1868A8eFB1182471F4De2121946Fd0"
+  }
+}
