+++ packages/adapters-library/src/adapters/pendle/products/standardised-yield-token/metadata/bsc.market.json
@@ -1,0 +1,87 @@
+{
+  "0x96F77F24b3d17166FcaF3a9E3B8330fF38B966F9": {
+    "protocolToken": {
+      "address": "0x96F77F24b3d17166FcaF3a9E3B8330fF38B966F9",
+      "name": "SY wbETH",
+      "symbol": "SY-wbETH",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0xa2E3356610840701BDf5611a53974510Ae27E2e1",
+        "name": "Wrapped Binance Beacon ETH",
+        "symbol": "wBETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x080f52A881ba96EEE2268682733C857c560e5dd4"
+  },
+  "0xf86DDA825885674F85C5420079E1Df1bC0F372c6": {
+    "protocolToken": {
+      "address": "0xf86DDA825885674F85C5420079E1Df1bC0F372c6",
+      "name": "SY Thena frxETH-ETH",
+      "symbol": "SY-Thena frxETH-ETH",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x8A420aACA0c92e3f97CDcFdd852e01Ac5b609452",
+        "name": "StableV1 AMM - ETH/frxETH",
+        "symbol": "sAMM-ETH/frxETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x66317Dec4a3d8d1e47b85f704E5DF675a68BB7C9"
+  },
+  "0x58732F1C1243A0D5876a601FB8D1Ee508B21580D": {
+    "protocolToken": {
+      "address": "0x58732F1C1243A0D5876a601FB8D1Ee508B21580D",
+      "name": "SY ankrBNB",
+      "symbol": "SY-ankrBNB",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x52F24a5e03aee338Da5fd9Df68D2b6FAe1178827",
+        "name": "Ankr Staked BNB",
+        "symbol": "ankrBNB",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0xEDa1D0e1681D59dea451702963d6287b844Cb94C"
+  },
+  "0xe49269B5D31299BcE407c8CcCf241274e9A93C9A": {
+    "protocolToken": {
+      "address": "0xe49269B5D31299BcE407c8CcCf241274e9A93C9A",
+      "name": "SY Renzo ezETH",
+      "symbol": "SY-ezETH",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x2416092f143378750bb29b79eD961ab195CcEea5",
+        "name": "Renzo Restaked ETH",
+        "symbol": "ezETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0xB2eEA27aF50030A445418553c4892065cf3a720a"
+  },
+  "0x56b70AC3a6Ae850F6AEE67591d1bB86895Ba211E": {
+    "protocolToken": {
+      "address": "0x56b70AC3a6Ae850F6AEE67591d1bB86895Ba211E",
+      "name": "SY weETH",
+      "symbol": "SY-weETH",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x04C0599Ae5A44757c0af6F9eC3b93da8976c150A",
+        "name": "Wrapped eETH",
+        "symbol": "weETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x27255F9aFF1868A8eFB1182471F4De2121946Fd0"
+  }
+}
