+++ packages/adapters-library/src/adapters/pendle/products/yield-token/metadata/bsc.market.json
@@ -1,0 +1,87 @@
+{
+  "0x429c77E193Ff7e6c6c3B2e1f00877B91792e87ee": {
+    "protocolToken": {
+      "address": "0x429c77E193Ff7e6c6c3B2e1f00877B91792e87ee",
+      "name": "YT wbETH 26DEC2024",
+      "symbol": "YT-wbETH-26DEC2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x96F77F24b3d17166FcaF3a9E3B8330fF38B966F9",
+        "name": "SY wbETH",
+        "symbol": "SY-wbETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x080f52A881ba96EEE2268682733C857c560e5dd4"
+  },
+  "0x8D0884EBA8E481CcADf2F91E62170cdF79AE9C64": {
+    "protocolToken": {
+      "address": "0x8D0884EBA8E481CcADf2F91E62170cdF79AE9C64",
+      "name": "YT Thena frxETH-ETH 27JUN2024",
+      "symbol": "YT-Thena frxETH-ETH-27JUN2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0xf86DDA825885674F85C5420079E1Df1bC0F372c6",
+        "name": "SY Thena frxETH-ETH",
+        "symbol": "SY-Thena frxETH-ETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x66317Dec4a3d8d1e47b85f704E5DF675a68BB7C9"
+  },
+  "0x269f3E230C90C0a29140106D4cFB5860aF6F6827": {
+    "protocolToken": {
+      "address": "0x269f3E230C90C0a29140106D4cFB5860aF6F6827",
+      "name": "YT ankrBNB 26JUN2025",
+      "symbol": "YT-ankrBNB-26JUN2025",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x58732F1C1243A0D5876a601FB8D1Ee508B21580D",
+        "name": "SY ankrBNB",
+        "symbol": "SY-ankrBNB",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0xEDa1D0e1681D59dea451702963d6287b844Cb94C"
+  },
+  "0x964B6D486a89025e5557a5aF3b4CC100A3F9734c": {
+    "protocolToken": {
+      "address": "0x964B6D486a89025e5557a5aF3b4CC100A3F9734c",
+      "name": "YT Renzo ezETH 26SEP2024",
+      "symbol": "YT-ezETH-26SEP2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0xe49269B5D31299BcE407c8CcCf241274e9A93C9A",
+        "name": "SY Renzo ezETH",
+        "symbol": "SY-ezETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0xB2eEA27aF50030A445418553c4892065cf3a720a"
+  },
+  "0xBf1E6E67cE2Bd8da1ea7342354126b59B46d4000": {
+    "protocolToken": {
+      "address": "0xBf1E6E67cE2Bd8da1ea7342354126b59B46d4000",
+      "name": "YT weETH 26SEP2024",
+      "symbol": "YT-weETH-26SEP2024",
+      "decimals": 18
+    },
+    "underlyingTokens": [
+      {
+        "address": "0x56b70AC3a6Ae850F6AEE67591d1bB86895Ba211E",
+        "name": "SY weETH",
+        "symbol": "SY-weETH",
+        "decimals": 18
+      }
+    ],
+    "marketAddress": "0x27255F9aFF1868A8eFB1182471F4De2121946Fd0"
+  }
+}
