+++ packages/adapters-library/src/adapters/pendle/tests/snapshots/ethereum.deposits.json
@@ -8,7 +8,7 @@
     "iconUrl": "https://app.pendle.finance/favicon.ico",
     "positionType": "supply",
     "chainId": 1,
-    "productId": "market",
+    "productId": "principle-token",
     "success": true,
     "movements": [
       {
