+++ packages/adapters-library/src/adapters/pendle/tests/snapshots/ethereum.profits.json
@@ -9,7 +9,7 @@
       "iconUrl": "https://app.pendle.finance/favicon.ico",
       "positionType": "supply",
       "chainId": 1,
-      "productId": "market",
+      "productId": "principle-token",
       "success": true,
       "tokens": [
         {
